requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
openpyxl
ollama
requests
